pyproject.toml
[build-system]
requires = [
    "setuptools>=40.8.0"
]
build-backend = "setuptools.build_meta"

[project]
name = "dohl"
version = "0.0.1"
authors = [
    { name = "Rte R&D and Eurobios SCB" },
]
description = "Damage in Overhead Lines application."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "matplotlib >= 3.4.3",
    "numpy >= 1.21.2",
    "pandas >= 1.4.1",
    "pyyaml >= 5.4.1",
    "scikit-learn >= 1.3.2",
]

[project.urls]
"Homepage" = "https://gitlab.eurobios.com/rte/dohl"
"Bug Tracker" = "https://fr.wikipedia.org/wiki/Erreur_HTTP_404"
